+++ modules/AWSPowerShell/Cmdlets/ElasticMapReduce/AWS.Tools.ElasticMapReduce.Completers.psm1
@@ -286,6 +286,7 @@ $EMR_SelectMap = @{
                "Get-EMRStudio",
                "Get-EMRAutoTerminationPolicy",
                "Get-EMRBlockPublicAccessConfiguration",
+               "Get-EMRClusterSessionCredential",
                "Get-EMRManagedScalingPolicy",
                "Get-EMRStudioSessionMapping",
                "Get-EMRBootstrapActionList",